21 Oct

+++ hourglass.cpp
@@ -5,7 +5,9 @@ using namespace std;
 
 int calculate_sum(vector <vector <int>> arr, int i, int j) {
     int sum;
-    sum = arr[i][j] +  arr[i][j + 1] +  arr[i][j + 2] +  arr[i + 1][j + 1] +  arr[i + 2][j] +  arr[i + 2][j + 1] +  arr[i + 2][j + 2];
+    sum = arr[i][j] +  arr[i][j + 1] +  arr[i][j + 2] + 
+     arr[i + 1][j + 1] +  
+     arr[i + 2][j] +  arr[i + 2][j + 1] +  arr[i + 2][j + 2];
     return sum;
 }
 
@@ -13,8 +15,8 @@ int largest_hourglass(vector <vector <int>> arr)
 {
     int max = -9 * 8;
     int sum;
-    for (uint i = 0; i < 4; i++)
-        for (uint j = 0; j < 4; j++) {
+    for (uint8_t i = 0; i < 4; i++)
+        for (uint8_t j = 0; j < 4; j++) {
             sum = calculate_sum(arr, i, j);
             max = max < sum? sum: max;
         }
